requirements.txt
streamlit==1.29.0
scikit-learn==1.3.0
numpy==1.26.2
matplotlib==3.8.0
pillow==10.0.1